agregado errores en update y detalle en producto no existente

+++ src/controllers/API/apiProductos.js
@@ -112,6 +112,10 @@ module.exports = {
             });
         } catch (error) {
             console.log(error);
+            return res.status(200).json({
+                ok: false,
+                errors:'Producto no encontrado'
+            });
         }
     },
     agregarProducto: async (req, res) => {
@@ -224,49 +228,59 @@ module.exports = {
                     },
                 });
     
+            
+                let product = await db.Product.findByPk(req.params.id, options(req));
+
                 if (!brandFound) {
                     brandFound = await db.Brand.create({
                         name: brand,
                     });
                 }
+                if (product){
+                    product.name = name.trim() || product.name;
+                    product.price = price || product.price;
+                    product.detail = detail.trim() || product.detail;
+                    product.amount = amount || product.amount;
+                    product.discount = discount || product.discount;
+                    product.category_id = categ.id || product.category_id;
+                    product.brand_id = brandFound.id || product.brand_id;
+                    await product.save();
 
-                let product = await db.Product.findByPk(req.params.id, options(req));
-
-                product.name = name.trim() || product.name;
-                product.price = price || product.price;
-                product.detail = detail.trim() || product.detail;
-                product.amount = amount || product.amount;
-                product.discount = discount || product.discount;
-                product.category_id = categ.id || product.category_id;
-                product.brand_id = brandFound.id || product.brand_id;
-                await product.save();
-
-                if(req.files && req.files.length){
-                    req.files.forEach(async (file, index) => {
+                    if(req.files && req.files.length){
+                    
+                        req.files.forEach(async (file, index) => {
                         
-                        if(product.imagenes[index]){
-                            fs.existsSync(path.join(__dirname,'..','..','..','public','images','productos',product.imagenes[index].file)) && fs.unlinkSync(path.join(__dirname,'..','..','..','public','images','productos',product.imagenes[index].file))
-                            product.imagenes[index].file = file.filename;
-                            product.imagenes[index].dataValues.url = `${req.protocol}://${req.get('host')}/productos/images/${file.filename}`
-                            await product.imagenes[index].save();
-                            await product.reload(options(req))
-                        }else{
-                            console.log(file.filename,path.join(__dirname,'..','..','..','public','images','productos',file.filename));
-                            await db.Image.create({
-                                product_id:product.id,
-                                file:file.filename,
-                                url:path.join(__dirname,'..','..','..','public','images','productos',file.filename)
-                            })
-                            await product.reload(options(req))
-                        }
+                            if(product.imagenes[index]){
+                                fs.existsSync(path.join(__dirname,'..','..','..','public','images','productos',product.imagenes[index].file)) && fs.unlinkSync(path.join(__dirname,'..','..','..','public','images','productos',product.imagenes[index].file))
+                                product.imagenes[index].file = file.filename;
+                                product.imagenes[index].dataValues.url = `${req.protocol}://${req.get('host')}/productos/images/${file.filename}`
+                                await product.imagenes[index].save();
+                                await product.reload(options(req))
+                            }else{
+                                console.log(file.filename,path.join(__dirname,'..','..','..','public','images','productos',file.filename));
+                                await db.Image.create({
+                                    product_id:product.id,
+                                    file:file.filename,
+                                    url:path.join(__dirname,'..','..','..','public','images','productos',file.filename)
+                                })
+                                await product.reload(options(req))
+                            }
+                        });
+                    }
+                    return res.status(201).json({
+                         ok : true,
+                        data : product,
                     });
-                }
-                
-                return res.status(201).json({
-                    ok : true,
-                    data : product,
-                });
-
+                }else{
+                    errorsDetail={
+                        ...errorsDetail,
+                        producto:'No existe el producto',
+                    }
+                    return res.status(200).json({
+                        ok: false,
+                        errors: errorsDetail,
+                    });
+                }             
 			}else{
                 if(req.files.length > 0){
                     req.files.forEach(({filename}) => {
